Hard Drive/backend/requirements.txt
anyio==4.8.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.1
chromadb==0.4.0
click==8.1.8
coloredlogs==15.0.1
einops==0.8.1
fastapi==0.99.1
filelock==3.17.0
Flask==2.1.1
flatbuffers==25.2.10
fsspec==2025.2.0
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
monotonic==1.6
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
onnxruntime==1.20.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
posthog==3.13.0
protobuf==5.29.3
pulsar-client==3.6.0
pydantic==1.10.21
PyPika==0.48.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.1.3
